requirements.txt
dotenv
fastapi[standard-no-fastapi-cloud-cli]
openai
pdf2image
Pillow
python-multipart
pydantic
pydantic[email]
pydantic_settings
sqlalchemy
uvicorn
